add: work

+++ philo.h
@@ -10,9 +10,9 @@
 
 typedef struct	s_data {
 	int			number_of_philosophers;
-	int			time_to_die;
-	int			time_to_eat;
-	int			time_to_sleep;
+	unsigned long long time_to_die;
+	unsigned long long time_to_eat;
+	unsigned long long time_to_sleep;
 	int			number_of_times_each_philosopher_must_eat;
 	int			full_philo;
 	pthread_mutex_t		print_message;
@@ -27,12 +27,12 @@ typedef struct		s_philosopher {
 	pthread_mutex_t	*left_fork;
 	pthread_mutex_t	*right_fork;
 	int				times_of_eating;
-	unsigned long	start_philo;
-	unsigned long	last_meal;
+	unsigned long long	start_philo;
+	unsigned long long	last_meal;
 }					t_philosopher;
 
 void*			life_cycle(void *philo);
-unsigned long	get_time(void);
+unsigned long long  get_time(void);
 int				ft_atoi(char *str);
 int				ft_isdigit(int c);
 void			print_status(t_philosopher *philosopher, int status);
